Refactor and clean serviceGroup component
