+++ src/client/src/containers/PageAddServices/ServiceCard/index.tsx
@@ -1,8 +1,8 @@
 import classnames from "classnames";
 import * as React from "react";
-
+import _ from "lodash";
 import { FormattedMessage, injectIntl, InjectedIntlProps } from "react-intl";
-
+import * as ModalActions from "../../../store/navigation/modals/action";
 
 import buttonStyles from "../../../css/buttonStyles.module.css";
 import styles from "./styles.module.css";
@@ -10,7 +10,12 @@ import { IOption } from "../../../types/option";
 import { getSvg } from "../../../utils/getSvgUrl";
 import CardBody from "../../../components/CardBody";
 import keyUpHandler from "../../../utils/keyUpHandler";
-import { KEY_EVENTS } from "../../../utils/constants";
+import { KEY_EVENTS, ROUTES, WIZARD_CONTENT_INTERNAL_NAMES } from "../../../utils/constants";
+import { useDispatch, useSelector } from "react-redux";
+import { setPageWizardPageAction, setDetailPageAction } from "../../../store/navigation/routes/action";
+import messages from "./messages";
+import { getServices } from "../../../store/userSelection/services/servicesSelector";
+import { isLoggedInSelector } from "../../../store/config/azure/selector";
 
 interface IProps {
   option: IOption;
@@ -21,30 +26,54 @@ type Props = IProps & InjectedIntlProps;
 export const ServiceCard = (props: Props) => {
   const { intl, option } = props;
   const { formatMessage } = intl;
+  const dispatch = useDispatch();
+  const isLoggedIn = useSelector(isLoggedInSelector);
+  const servicesSelection = useSelector(getServices);
 
-/*
+  const handleDetailsClick = (option: IOption) => {
+    const isIntlFormatted = true;
+    dispatch(setPageWizardPageAction(ROUTES.PAGE_DETAILS));
+    dispatch(setDetailPageAction(option, isIntlFormatted, ROUTES.ADD_SERVICES));
+  };
+
+  const isSelectionCreated = (internalName: string): boolean => {
+    if (internalName === WIZARD_CONTENT_INTERNAL_NAMES.COSMOS_DB) {
+      return !_.isEmpty(servicesSelection.cosmosDB);
+    } else if (internalName === WIZARD_CONTENT_INTERNAL_NAMES.APP_SERVICE) {
+      return !_.isEmpty(servicesSelection.appService);
+    }
+    return false;
+  };
+
   const addOrEditResourceText = (internalName: string): string => {
     if (isSelectionCreated(internalName)) {
       return formatMessage(messages.editResource);
     }
-    return formatMessage(messages.addResource);
-  };
-*/
-
-  const handleDetailsClick = (option: IOption) => {
-    console.log(option);
+    return formatMessage(messages.addToProject);
   };
   
-  const handleButtonClick = (option: IOption) => {
-    console.log(option);
-  };
-
   const handleDetailsClickIfPressEnterKey = (event: React.KeyboardEvent<HTMLAnchorElement>, option: IOption) => {
     event.stopPropagation();
     if (event.key === KEY_EVENTS.ENTER) {
       handleDetailsClick(option);
     }
   };
+
+  const openCloudServiceModal = (option: IOption) => {
+    const modalOpeners = {
+      [WIZARD_CONTENT_INTERNAL_NAMES.COSMOS_DB]: () => dispatch(ModalActions.openCosmosDbModalAction()),
+      [WIZARD_CONTENT_INTERNAL_NAMES.APP_SERVICE]: ()=> dispatch(ModalActions.openAppServiceModalAction())
+    };
+    if (modalOpeners.hasOwnProperty(option.internalName)) {
+      modalOpeners[option.internalName]();
+    }
+    //return () => void(0);
+  }
+
+  const openLoginModal = (option: IOption) => {
+    dispatch(ModalActions.openAzureLoginModalAction(option.internalName));
+  }
+
   return (
     <div className={styles.container}>
     <div className={styles.loginContainer}>
@@ -71,7 +100,11 @@ export const ServiceCard = (props: Props) => {
             <FormattedMessage id="card.details" defaultMessage="Learn more" />
           </a>
           <button
-            onClick={() => handleButtonClick}
+            onClick={
+              isLoggedIn
+                ? () => openCloudServiceModal(option)
+                : () => openLoginModal(option)
+            }
             className={classnames(
               styles.signInButton,
               buttonStyles.buttonHighlighted,
@@ -79,7 +112,7 @@ export const ServiceCard = (props: Props) => {
             )}
             tabIndex={0}
           >
-            {"VER addOrEditResourceText"}
+            {addOrEditResourceText(option.internalName)}
           </button>
         </div>
       </div>
